[Refactor] Split glog init from runtime logconfig
