+++ be/src/exec/pipeline/exchange/multi_cast_local_exchange.cpp
@@ -16,9 +16,9 @@
 
 #include <glog/logging.h>
 
+#include "base/logging.h"
 #include "exec/pipeline/exchange/multi_cast_local_exchange_sink_operator.h"
 #include "runtime/runtime_state.h"
-#include "util/logging.h"
 
 namespace starrocks::pipeline {
