+++ be/test/storage/persistent_index_consistency_test.cpp
@@ -21,6 +21,7 @@
 #include <random>
 
 #include "base/coding.h"
+#include "base/logging.h"
 #include "base/string/faststring.h"
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
@@ -38,7 +39,6 @@
 #include "storage/tablet_manager.h"
 #include "storage/update_manager.h"
 #include "testutil/deterministic_test_utils.h"
-#include "util/logging.h"
 
 DEFINE_bool(debug, false, "debug mode");
 DEFINE_int64(seed, -1, "random seed");
