+++ be/test/testutil/deterministic_test_utils.h
@@ -22,10 +22,10 @@
 #include <random>
 
 #include "base/coding.h"
+#include "base/logging.h"
 #include "base/string/faststring.h"
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
-#include "util/logging.h"
 
 namespace starrocks {
