+++ be/test/testutil/init_test_env.h
@@ -23,6 +23,7 @@
 #include "common/config_cache_fwd.h"
 #include "common/config_path_fwd.h"
 #include "common/config_storage_fwd.h"
+#include "common/glog_init.h"
 #include "common/metrics/process_metrics_registry.h"
 #include "common/system/cpu_info.h"
 #include "common/system/disk_info.h"
@@ -43,7 +44,6 @@
 #include "storage/tablet_manager.h"
 #include "storage/update_manager.h"
 #include "types/time_types.h"
-#include "util/logging.h"
 
 namespace starrocks {
